Database now fully working for persistence.
Probably not the safest design methodology. As the hard storage in the database is wiped onPause() and is replaced by whatever is in the cache. While it does remove duplicates, it may present other problems down the line, in addition to also being database intensive if we have a large number of Tasks.